Create all_code
This is all code used to analyse and visualise the data for Foister & Wilson, 'Anthropocentrism as a source of sampling bias in the fossil record'. The data used is an export of the NOW database (https://nowdatabase.luomus.fi/index.php). Not all plots produced by this code are used in the final manuscript as they were not useful visuals for readers, but they aided in our analysis.